rebased on year-view, formatted, added date-helpers file to lib, deleted functions folder and added dark mode border color
