+++ src/components/calendar-header.tsx
@@ -1,3 +1,5 @@
 export default function CalendarHeader() {
-	return <div className="flex items-center border p-5 m-5 rounded-t-2xl" />;
+  return (
+    <div className="flex items-center border dark:border-gray-200 p-5 border-b-0 rounded-t-2xl" />
+  );
 }
