+++ src/components/year-view.tsx
@@ -1,5 +1,9 @@
+import MonthContainers from "@/components/year-view-components/month-containers.tsx";
+
 export default function YearView() {
   return (
-    <div className=" m-4 p-4 rounded-b-2xl border border-gray-300 flex  justify-center " />
+    <div className="p-4 rounded-b-2xl border dark:border-gray-200 flex  justify-center ">
+      <MonthContainers />
+    </div>
   );
 }
